# Udaconnect-message-passing/docs/openAPI.yaml
swagger: '2.0'
basePath: /
paths:
  /api/locations:
    parameters:
      - in: query
        description: (required) Unique ID for a given Location
        name: location_id
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_location_resource
      tags:
        - Udaconnect
    post:
      responses:
        '200':
          description: Success
      operationId: post_location_resource
      tags:
        - Udaconnect
  '/api/locations/{location_id}':
    parameters:
      - in: path
        description: Unique ID for a given Location
        name: location_id
        required: true
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_locations_resource
      tags:
        - Udaconnect
    post:
      responses:
        '200':
          description: Success
      operationId: post_locations_resource
      tags:
        - Udaconnect
  /api/persons:
    post:
      responses:
        '200':
          description: Success
      operationId: post_persons_resource
      tags:
        - Udaconnect
  '/api/persons/{person_id}':
    parameters:
      - in: path
        description: Unique ID for a given Person
        name: person_id
        required: true
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_person_resource
      tags:
        - Udaconnect
  '/api/persons/{person_id}/connection':
    parameters:
      - in: query
        description: Proximity to a given user in meters
        name: distance
        type: string
      - in: query
        description: Upper bound of date range
        name: end_date
        type: string
      - in: query
        description: Lower bound of date range
        name: start_date
        type: string
      - name: person_id
        in: path
        required: true
        type: string
    get:
      responses:
        '200':
          description: Success
      operationId: get_connection_data_resource
    post:
      responses:
        '200':
          description: Success
      operationId: post_connection_data_resource
      tags:
      tags:
        - Udaconnect
info:
  title: Udaconnect API
  version: 0.1.0
produces:
  - application/json
consumes:
  - application/json
tags:
  - name: Udaconnect
    description: Connections via geolocation.
